course_mcp/2. MCP with API/1. MCP with GitHub/requirements.txt
fastmcp==3.1.1
mcp==1.26.0
httpx==0.28.1
python-dotenv==1.2.2
